inline SnapstartHelperLambdaRequest
